% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_simulate_standings.R
\name{do_simulate_standings}
\alias{do_simulate_standings}
\title{Combine \code{generate_schedules} and \code{scrape_espn_ff_scores} results}
\usage{
do_simulate_standings(
  league_id = .get_league_id(),
  league_size = .get_league_size(),
  season = .get_season(),
  weeks = .get_weeks_cutoff(),
  sims = 10,
  tries = ceiling(log(sims)),
  overwrite = FALSE,
  export = TRUE,
  dir = .get_dir_data(),
  file = .generate_sims_file(league_id, league_size, season, weeks, sims),
  ext = "parquet",
  path = .generate_path(dir, file, ext),
  f_import = arrow::read_parquet,
  f_export = arrow::write_parquet,
  params.schedules = list(),
  params.scores = list()
)
}
\arguments{
\item{league_id}{Number for ESPN league. Probably 6 digits. Can be set globally
in the options. See \code{ffsched.league_id}.}

\item{league_size}{Number of teams in the league. Can be set globally
in the options. See \code{ffsched.league_size}.}

\item{season}{Season for which to scrape. Can be set globally in the options.
See \code{ffsched.season}.}

\item{weeks}{How many weeks are in schedule. Presently, this function requires
that \verb{league_size <= league_size < (2 * (league_size - 1))}.}

\item{sims}{How many unique simulations to generate.}

\item{tries}{How many times to re-try.}

\item{overwrite}{Whether to overwrite existing file at \code{path}, if it exists.}

\item{export}{Whether to export.}

\item{dir}{Directory to use to generate \code{path} if \code{path} is not explicitly provided.}

\item{file}{Filename (without extension) to generate \code{path} if \code{path} is not explicitly provided.}

\item{ext}{File extension to use to generate \code{path} if \code{path} is not explicitly provided.}

\item{path}{Path to export to.}

\item{f_import}{Function to import with if file exists and \code{overwrite = TRUE}.}

\item{f_export}{Function to export with if \code{export = TRUE} .}

\item{params.schedules, params.scores}{List of named arguments that share
names with arguments intended for this function (e.g. \code{ext}, \code{file}, \code{path}, etc.)
that should be specific to the \code{generate_schedules} and \code{scrape_espn_ff_scores}.
You probably don't want to use these arguments, unless you have a speficic
use case (e.g re-run only 1 of the functions or name the output file
in a certain, custom way).}
}
\description{
Combine \code{generate_schedules} and \code{scrape_espn_ff_scores} results to come up
with final standing frequencies
}
