% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_schedules.R
\name{generate_schedule}
\alias{generate_schedule}
\title{Generate a single schedule}
\usage{
generate_schedule(
  league_size = .get_league_size(),
  seed = NULL,
  seed_init = NULL,
  retries = league_size^2,
  retries_init = retries
)
}
\arguments{
\item{league_size}{Number of teams in the league. Can be set globally
in the options. See \code{ffsched.league_size}.}

\item{seed}{Seed to use}

\item{seed_init}{Initial seed to use, passed to the non-exported \code{.initialize()} function.}

\item{retries}{Number of retries for a single schedule}

\item{retries_init}{Initial number of retries, passed to the non-exported \code{.initialize()} function.}
}
\description{
Generate a single schedule
}
\details{
Iterate by week and team, always checking for other matched teams in
the same week and previously played teams in prior weeks
}
