% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape_espn_ff_scores}
\alias{scrape_espn_ff_scores}
\title{Scrape ESPN fantasy football scores}
\usage{
scrape_espn_ff_scores(
  league_id = .get_league_id(),
  league_size = .get_league_size()
)
}
\arguments{
\item{league_id}{Number for ESPN league. Probably 6 digits. Can be set globally
in the options. See \code{ff.league_id}.}

\item{league_size}{Number of teams in the league. Can be set globally
in the options. See \code{ff.league_size}.}
}
\description{
Scrape ESPN fantasy football scores
}
