% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_set_update_ffsched}
\alias{theme_set_update_ffsched}
\title{Theme for ggplots}
\usage{
theme_set_update_ffsched(...)
}
\arguments{
\item{...}{Extra arguments to pass to \code{ggplot2::theme_update}}
}
\description{
Theme for ggplots
}
