% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_schedules.R
\name{generate_schedules}
\alias{generate_schedules}
\title{Generate schedules}
\usage{
generate_schedules(
  league_size = .get_league_size(),
  weeks = .get_weeks_cutoff(),
  sims = 10,
  tries = ceiling(log(sims)),
  check_dups = TRUE,
  ...,
  overwrite = FALSE,
  export = TRUE,
  dir = .get_dir_out(),
  file = .generate_schedules_file(league_size, weeks, sims),
  ext = "parquet",
  path = file.path(dir, sprintf("\%s.\%s", file, ext)),
  f_import = arrow::read_parquet,
  f_export = arrow::write_parquet
)
}
\arguments{
\item{league_size}{Number of teams in the league. Can be set globally
in the options. See \code{ff.league_size}.}

\item{weeks}{How many weeks are in schedule. Presently, this function requires
that \verb{league_size <= league_size < (2 * (league_size - 1))}.}

\item{sims}{How many unique simulations to generate.}

\item{tries}{How many times to re-try.}

\item{check_dups}{Whether to check for duplicates. It's recommended to leave this
as \code{TRUE} to ensure that you get unique results, but that may not be what you
desire.}

\item{...}{Additional parameters passed to \code{generate_schedule()}}

\item{overwrite}{Whether to overwrite existing file at \code{path}, if it exists.}

\item{export}{Whether to export.}

\item{dir, file, ext}{Directory, filename (without extension), and file extension
to use to generate \code{path} if \code{path} is not explicitly provided.}

\item{path}{Path to export to.}

\item{f_import}{Function to import with if file exists and \code{overwrite = TRUE}.}

\item{f_export}{Function to export with if \code{export = TRUE} .}
}
\description{
Generate unique schedules for teams, presumably for a fantasy football league,
but it doesn't actually have to be for that. Points are not simulated, just
schedules
}
\details{
This function basically uses a brute force approach, hence the \code{tries}
parameter.
}
