% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_espn_ff_scores.R
\name{scrape_espn_ff_scores}
\alias{scrape_espn_ff_scores}
\title{Scrape ESPN fantasy football scores}
\usage{
scrape_espn_ff_scores(
  league_id = .get_league_id(),
  league_size = .get_league_size(),
  season = .get_season(),
  ...,
  overwrite = FALSE,
  export = TRUE,
  dir = .get_dir_out(),
  file = .generate_espn_ff_scores_file(league_id, league_size, season),
  ext = "csv",
  path = file.path(dir, sprintf("\%s.\%s", file, ext)),
  f_import = readr::read_csv,
  f_export = readr::write_csv
)
}
\arguments{
\item{league_id}{Number for ESPN league. Probably 6 digits. Can be set globally
in the options. See \code{ff.league_id}.}

\item{league_size}{Number of teams in the league. Can be set globally
in the options. See \code{ff.league_size}.}

\item{season}{Season for which to scrape. Can be set globally in the options.
See \code{ff.season}.}

\item{...}{Additional parameters passed to \code{generate_schedule()}}

\item{overwrite}{Whether to overwrite existing file at \code{path}, if it exists.}

\item{export}{Whether to export.}

\item{dir}{Directory, filename (without extension), and file extension
to use to generate \code{path} if \code{path} is not explicitly provided.}

\item{file}{Directory, filename (without extension), and file extension
to use to generate \code{path} if \code{path} is not explicitly provided.}

\item{ext}{Directory, filename (without extension), and file extension
to use to generate \code{path} if \code{path} is not explicitly provided.}

\item{path}{Path to export to.}

\item{f_import}{Function to import with if file exists and \code{overwrite = TRUE}.}

\item{f_export}{Function to export with if \code{export = TRUE} .}
}
\description{
Scrape ESPN fantasy football scores
}
\seealso{
\url{https://gist.github.com/lbenz730/ea7d5bce0a36fe66c4241c8facd6c153}
}
