% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_espn_ff_scores.R
\name{scrape_espn_ff_scores}
\alias{scrape_espn_ff_scores}
\title{Scrape ESPN fantasy football scores}
\usage{
scrape_espn_ff_scores(
  league_id = .get_league_id(),
  league_size = .get_league_size(),
  season = .get_season(),
  weeks = .get_weeks_cutoff(),
  local = FALSE,
  path_teams = NULL,
  path_scores = NULL,
  ...,
  overwrite = FALSE,
  export = TRUE,
  dir = .get_dir_data(),
  file = .generate_espn_ff_scores_file(league_id, league_size, season, weeks),
  ext = "csv",
  path = file.path(dir, sprintf("\%s.\%s", file, ext)),
  f_import = readr::read_csv,
  f_export = readr::write_csv
)
}
\arguments{
\item{league_id}{Number for ESPN league. Probably 6 digits. Can be set globally
in the options. See \code{ffsched.league_id}.}

\item{league_size}{Number of teams in the league. Can be set globally
in the options. See \code{ffsched.league_size}.}

\item{season}{Season for which to scrape. Can be set globally in the options.
See \code{ffsched.season}.}

\item{weeks}{How many weeks are in schedule. Presently, this function requires
that \verb{league_size <= league_size < (2 * (league_size - 1))}.}

\item{local, path_teams, path_scores}{If your league is not public, you will have
to download the appropriate JSON files, set \code{local = TRUE},
and point \code{path_teams} and \code{path_scores} to these JSON files.
The JSON for \code{path_teams} should downloaded from
\url{http://fantasy.espn.com/apis/v3/games/ffl/seasons/{season}/segments/0/leagues/{league_id}?view=mtm}
and the JSON for \code{path_scores} should be downloaded from
\url{http://fantasy.espn.com/apis/v3/games/ffl/seasons/{season}/segments/0/leagues/{league_id}?view=mMatchup}.}

\item{...}{Additional parameters passed to \code{generate_schedule()}}

\item{overwrite}{Whether to overwrite existing file at \code{path}, if it exists.}

\item{export}{Whether to export.}

\item{dir}{Directory to use to generate \code{path} if \code{path} is not explicitly provided.}

\item{file}{Filename (without extension) to generate \code{path} if \code{path} is not explicitly provided.}

\item{ext}{File extension to use to generate \code{path} if \code{path} is not explicitly provided.}

\item{path}{Path to export to.}

\item{f_import}{Function to import with if file exists and \code{overwrite = TRUE}.}

\item{f_export}{Function to export with if \code{export = TRUE} .}
}
\description{
Scrape ESPN fantasy football scores
}
\seealso{
\url{https://gist.github.com/lbenz730/ea7d5bce0a36fe66c4241c8facd6c153}
}
